Moved BorrowedObjectPool to common
