+++ base/common/BorrowedObjectPool.h
@@ -7,8 +7,7 @@
 #include <condition_variable>
 
 #include <common/defines.h>
-
-#include <Common/MoveOrCopyIfThrow.h>
+#include <common/MoveOrCopyIfThrow.h>
 
 /** Pool for limited size objects that cannot be used from different threads simultaneously.
   * The main use case is to have fixed size of objects that can be reused in difference threads during their lifetime
